refactoring of navigation
