+++ apps/chat/src/hooks/useWidgets.ts
@@ -2,6 +2,7 @@ import { useCallback, useEffect, useMemo } from 'react';
 
 import { useRouter } from 'next/router';
 
+import { WidgetsSelectors } from '../store/widgets/widgets.selectors';
 import { ApplicationActions } from '@/src/store/actions';
 import { ApplicationSelectors } from '@/src/store/application/application.selectors';
 import { useAppDispatch, useAppSelector } from '@/src/store/hooks';
@@ -13,18 +14,7 @@ import { Routes } from '../constants/routes';
 export const useWidgets = () => {
   const router = useRouter();
 
-  const widgetsSchemaIds = useAppSelector(
-    SettingsSelectors.selectWidgetsSchemaIds,
-  );
-  const models = useAppSelector(ModelsSelectors.selectModels);
-
-  const widgetModels = useMemo(() => {
-    return models
-      .filter((m) => widgetsSchemaIds.has(m.applicationTypeSchemaId ?? ''))
-      .toSorted((a, b) =>
-        a.name.localeCompare(b.name, undefined, { sensitivity: 'base' }),
-      );
-  }, [models, widgetsSchemaIds]);
+  const widgetModels = useAppSelector(WidgetsSelectors.selectWidgets);
 
   const handleWidgetClick = useCallback(
     (id: string) => {
